DesignProject/requirements.txt
opencv-python==4.8.1.78
tensorflow==2.13.0
numpy==1.24.3
scikit-learn==1.3.0
pyserial==3.5
pyrealsense2